Fix for bug #138 - Add map picker to android app
For storing dive location, a map picker is added.
One could choose location of dive as well as time offset in minutes
of the dive and store the information on server or locally
as per preferences.
For doing this, click the 'plus' symbol in the action bar and chose
'Use map' option. Pick location on map and press 'send' icon on the
actionbar. Chose name of the dive and time offset in minutes.
Pressing okay will save locally or send to server as per your
preference.
Changes made:
1. Added map picker option on add button in actionbar.
	a. Modified HomeActivity.java activity and dives.xml menu
	b. Added PickLocation.java activity and dive_picklocation.xml
	   layout and dive_loc_menu.xml menu
2. The new strings are included in languages German, French, Russian
   and Romanian.
	a. Added strings in strings.xml of each language
3. Changed if else ladder in HomeActivity.java to switch case
   for better code readability
        a.Modified HomeActivity.java

Reported-by: henrik
Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>