Case of absence of sym and/or name tag in gpx file
In case the sym and/or name tags are not present in the GPX file
appropriate measures are taken.
In case neither are present a default string is chosen.
Made suitable changes such that string localisation is possible.
Also, if the divetime is absent in the GPX, the present device
time is taken as the divetime.
Previously as the timestamp was set to -1, the import showed improper
result.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>